+++ common/src/main/java/com/chefmooon/colourfulclocks/common/item/BornholmTopBlockItem.java
@@ -1,7 +1,10 @@
 package com.chefmooon.colourfulclocks.common.item;
 
 import com.chefmooon.colourfulclocks.common.data.BornholmTopGlassComponent;
+import com.chefmooon.colourfulclocks.common.data.types.BornholmTopGlassTypes;
+import com.chefmooon.colourfulclocks.common.data.types.PocketWatchTypes;
 import com.chefmooon.colourfulclocks.common.registry.ColourfulClocksDataComponentTypes;
+import com.chefmooon.colourfulclocks.common.util.ColourfulClocksTypeUtil;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.*;
 import net.minecraft.world.level.block.Block;
@@ -18,14 +21,13 @@ public class BornholmTopBlockItem extends BlockItem {
     public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
         if (stack.has(ColourfulClocksDataComponentTypes.getBornholmTopGlassData())) {
             Item glassItem = stack.getOrDefault(ColourfulClocksDataComponentTypes.getBornholmTopGlassData(), BornholmTopGlassComponent.getDefaultValue()).getGlassType().getItem();
-            if (glassItem != Items.AIR) {
+            if (glassItem != BornholmTopGlassTypes.GLASS.getItem()) {
                 tooltipComponents.add(Component.translatable(glassItem.getDescriptionId()));
             }
-            // TODO : figure out how to add pocket watch from data on place
-//            String dial = stack.getOrDefault(ColourfulClocksDataComponentTypes.getBornholmTopGlassData(), BornholmTopGlassComponent.getDefaultValue()).getPocketWatchType().getBaseTranslation();
-//            if (!dial.isEmpty()) {
-//                tooltipComponents.add(Component.translatable("Dial: " + dial));
-//            }
+            Item pocketWatchItem = ColourfulClocksTypeUtil.getPocketWatchItemFromType(stack.getOrDefault(ColourfulClocksDataComponentTypes.getBornholmTopGlassData(), BornholmTopGlassComponent.getDefaultValue()).getPocketWatchType());
+            if (pocketWatchItem != PocketWatchTypes.EMPTY.getItem()) {
+                tooltipComponents.add(Component.translatable(pocketWatchItem.getDescriptionId()));
+            }
         }
 
         super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);